Project/03_coffee_shop_full_stack/starter_code/backend/requirements.txt
astroid==3.0.1
Click==8.1.3
ecdsa==0.18.0
Flask==2.2.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.3
future==0.18.3
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.1
mccabe==0.7.0
pycryptodome==3.3.1
python-dotenv==1.0.0
pylint==3.0.2
python-jose==3.3.0
python-jose-cryptodome==1.3.2
six==1.16.0
typed-ast==1.5.4
Werkzeug==2.2.2
wrapt==1.15.0